Switch to X east, Z north coordinates
The memory layout of Harry and the camera's coordinates (increasing the first int moves east) suggests standard mnemonic XYZ ordering. Furthermore, the orientation of HERO.ILM when it's loaded shows that meshes are built facing north. I also finally realized that the game's Euler rotations are relative to objects' local space, and on top of that I figured out how to properly sync up the cameras, so my previous rationalizations are moot.