Clip polygons properly
Finally! My corners are filled in as intended, and all is right with the world. The option to disable frustum culling made no sense (leads to crashes with way-offscreen coordinates) and is gone. Performance is also in the toilet, but step one was making it work correctly, next up is doing it fast.