Read latest commit message as string, not array
This is what was wrong the whole time! A one-line commit message was treated as a string, but multiple lines became an array, which complicated both the StartsWith call and use of the -match command. Piping through the Out-String commandlet should do the trick.

So [skip benchmarks] already, damn it.