Set up .NET on D: instead of the OS drive
Previous slowness I'd seen during the Setup .NET step wasn't the download, which was only 5 or 6 seconds, but the .zip extraction. According to what I've read since then, hosted Windows runners for now have an issue with the C: drive being slower than D:, and the suggestion was to install in the faster location as a quick fix. Let's see how that works out.