Optimize benchmark skip Powershell script
Merge is the first word of this message as a test. Checking for [skip benchmarks] first (yes I just typed it, that's another test) avoids the rare, but still possible, situation of a merge commit in a PR that's not one of Github's invisible ones after a force push. If the text is there, it's important, so obey it. Checking for 'Merge' afterwards ultimately gets the same result as checking beforehand.

This commit also gets rid of one unnecessary variable and hoists another out of the loop. Vanishingly minor imperfections, but they still bugged me.