Require ContinuousIntegrationBuild for packaging
Although the DeterministicSourcePaths property can be successfully updated during the MSBuild execution phase, I assume it getting set is not the only impact of ContinuousIntegrationBuild. Why would the two properties be separate if they're synonymous? Since the latter only takes effect during the evaluation phase of a build, just leave it for users to specify at the command line and throw an error if they forget, like I will a matter of weeks from now.

[skip ci]