Only force camera yaw in first person
There was a reason I originally put this null check outside of the check for first person mode, but I can't remember what it was. Presumably I mistakenly thought that was the solution to some problem that no longer exists. Let's hope so, because doing so broke going through doors in third person.