Preserve user settings on version change
Settings that still exist are kept as-is, anything obsolete is removed, new stuff is set to its default, and the file's version is simply bumped to current. Only in cases of actual load failure will existing configs be renamed and replaced by all defaults. As for renamed settings, there's no way to map old to new as far as I know, but that will hopefully be rare.