Use BinaryPrimitives for SH type byte conversion
Slightly lower performance, but still fast enough (getting a type's bytes should be rare anyway) while properly accommodating big endian architecture. The rest of the project doesn't yet, but one thing at a time. Also the BP methods are vastly improved in .NET 8, i.e. as fast as or faster than what I had before, so when the time comes for that we'll already be in business.