Don't carry original entity bytes around
They can easily be loaded directly from the ROM when needed. Ideally it will eventually be possible to persist customizations somehow, requiring 3 versions of an object (the disc original, the saved user customization, and the current customization that hasn't been saved yet) but we'll jump off that bridge when we come to it.